+++ src/client/gl-buffer.h
@@ -19,13 +19,13 @@ class GlBuffer final : public IResource {
 
   /** Used in the update thread */
   void GlBufferData(
-      const void *data, uint64_t target, size_t size, uint64_t usage);
+      const void *data, uint32_t target, size_t size, uint32_t usage);
 
   uint64_t id() override;
 
   /** Used in the rendering thread */
   inline GLuint buffer_id();
-  inline uint64_t target();
+  inline uint32_t target();
 
  private:
   const uint64_t id_;
@@ -36,8 +36,8 @@ class GlBuffer final : public IResource {
     size_t size = 0;
     size_t alloc = 0;
 
-    uint64_t target;
-    uint64_t usage;
+    uint32_t target;
+    uint32_t usage;
 
     bool data_damaged = false;
   } pending_;
@@ -45,7 +45,7 @@ class GlBuffer final : public IResource {
   struct {
     GLuint buffer_id = 0;
 
-    uint64_t target;
+    uint32_t target;
   } rendering_;
 };
 
@@ -55,7 +55,7 @@ GlBuffer::buffer_id()
   return rendering_.buffer_id;
 }
 
-inline uint64_t
+inline uint32_t
 GlBuffer::target()
 {
   return rendering_.target;
